+++ src/lib/kotlin/slatekit-actors/src/main/kotlin/slatekit/actors/Context.kt
@@ -2,6 +2,7 @@ package slatekit.actors
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import slatekit.actors.pause.Scheduler
 
 /**
  * Contains current ( and future ) dependencies for an actor
