+++ src/lib/kotlin/slatekit-actors/src/main/kotlin/slatekit/actors/pause/Check.kt
@@ -1,4 +1,4 @@
-package slatekit.actors
+package slatekit.actors.pause
 
 import slatekit.actors.Status
