+++ src/lib/kotlin/slatekit-actors/src/main/kotlin/slatekit/actors/pause/Controls.kt
@@ -1,4 +1,7 @@
-package slatekit.actors
+package slatekit.actors.pause
+
+import slatekit.actors.Action
+import slatekit.actors.Reference
 
 
 /**
