+++ src/lib/kotlin/slatekit-actors/src/main/kotlin/slatekit/actors/pause/Feedback.kt
@@ -1,3 +1,3 @@
-package slatekit.actors
+package slatekit.actors.pause
 
 data class Feedback(val success:Boolean, val msg:String)
