+++ src/lib/kotlin/slatekit-integration/src/main/kotlin/slatekit/integration/apis/JobsApi.kt
@@ -13,18 +13,16 @@
 
 package slatekit.integration.apis
 
-import slatekit.actors.Feedback
+import slatekit.actors.pause.Feedback
 import slatekit.apis.Api
 import slatekit.apis.Action
 import slatekit.apis.AuthModes
 import slatekit.apis.Verbs
 import slatekit.apis.support.FileSupport
-import slatekit.common.Identity
 import slatekit.common.Sources
 import slatekit.common.crypto.Encryptor
 import slatekit.common.log.Logger
 import slatekit.integration.common.AppEntContext
-import slatekit.jobs.Job
 import slatekit.jobs.Jobs
 import slatekit.jobs.support.Command
 import slatekit.results.Outcome
