+++ src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/Worker.kt
@@ -3,7 +3,7 @@ package slatekit.jobs
 import java.util.concurrent.atomic.AtomicReference
 import slatekit.common.Identity
 import slatekit.actors.Status
-import slatekit.actors.Check
+import slatekit.actors.pause.Check
 
 
 /**
